edited product quantity

+++ src/main/java/uz/pdp/springsecurity/entity/ExchangeProduct.java
@@ -17,7 +17,7 @@ import javax.persistence.OneToOne;
 @Entity
 public class ExchangeProduct  extends AbsEntity {
 
-    private Integer exchangeProductQuantity;
+    private Double exchangeProductQuantity;
 
     @OneToOne
     @OnDelete(action = OnDeleteAction.CASCADE)
